requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
astunparse==1.6.3
attrs==25.4.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
comtypes==1.4.13
contourpy==1.3.3
cycler==0.12.1
flatbuffers==25.9.23
fonttools==4.60.1
gast==0.6.0
google-ai-generativelanguage==0.6.15
google-api-core==2.26.0
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.62.3
h5py==3.15.0
httplib2==0.31.0
idna==3.11
jax==0.7.1
jaxlib==0.7.1
joblib==1.5.2
keras==3.11.3
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
mediapipe==0.10.21
ml_dtypes==0.5.3
namex==0.1.0
numpy==2.2.6
opencv-contrib-python==4.11.0.86
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.1
pydantic_core==2.41.3
Pygments==2.19.2
pyparsing==3.2.5
pypiwin32==223
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pyttsx3==2.99
pytz==2025.2
pywin32==311
requests==2.32.5
rich==14.2.0
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.2
sentencepiece==0.2.1
six==1.17.0
sounddevice==0.5.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.17.3
